# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.UserLogin:
    properties:
      blockOrNot:
        type: boolean
      email:
        type: string
      password:
        type: string
    type: object
...
  controllers.data:
    properties:
      area:
        type: string
      city:
        type: string
      defaultAdd:
        type: boolean
      district:
        type: string
      email:
        type: string
      first_name:
        type: string
      fullName:
        type: string
      houseName:
        type: string
      is_admin:
        type: boolean
      landmark:
        type: string
      last_name:
        type: string
      phone:
        type: string
      phoneAddrs:
        type: integer
      pincode:
        type: integer
      state:
        type: string
    type: object
...
  controllers.userData:
    properties:
      email:
        type: string
      password:
        type: string
      repeatePassword:
        type: string
    type: object
...
  models.User:
    properties:
      blockOrNot:
        type: boolean
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      otp:
        type: string
      password:
        type: string
      phone:
        type: string
      referal_code:
        type: string
      user_status:
        type: boolean
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
host: localhost:1111
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: An e-commerce site in Go using Gin framework
  title: E-commerce
  version: "1.0"
paths:
  /admin/usermanagement/manageblock:
    patch:
      description: Block or unblock a user by their user ID
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      responses:
        "200":
          description: User status updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      summary: Block or Unblock users
      tags:
      - Admin
  /admin/usermanagement/viewusers:
    get:
      consumes:
      - '*/*'
      description: Retrieve a list of active users
      produces:
      - application/json
      responses:
        "200":
          description: List of active users
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      summary: List Users
      tags:
      - Admin
...
  /user/editprofile:
    patch:
      consumes:
      - application/json
      description: Edit user profile details and associated address
      parameters:
      - description: User profile data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.data'
      produces:
      - application/json
      responses:
        "200":
          description: Updated profile details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Edit User Profile
      tags:
      - Users
...
  /user/forgotpassword:
    patch:
      consumes:
      - application/json
      description: User can change password if forgot
      parameters:
      - description: User input
        in: body
        name: user
        schema:
          $ref: '#/definitions/controllers.userData'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Forgot Password
      tags:
      - Authentication
      - Users
...
  /user/login:
    post:
      consumes:
      - '*/*'
      description: Log in a user and generate an authentication token
      parameters:
      - description: User login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: User Login
      tags:
      - Authentication
...
  /user/signup:
    post:
      consumes:
      - '*/*'
      description: Signup a user with required datas
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Successful registration
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User Signup
      tags:
      - Signup
  /user/viewprofile:
    get:
      description: Retrieve user details and asscciated address
      produces:
      - application/json
      responses:
        "202":
          description: User details and address
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Show User Details
      tags:
      - Users
schemes:
- http
swagger: "2.0"